Silently tries to remove an  existing game on the lobbyserver, supports 64bit ints for gameId.

+++ lib/netplay/netlobby.h
@@ -127,7 +127,7 @@ class LobbyClient {
 		}
 
 	private:
-		int32_t gameId_;
+		int64_t gameId_;
 
 		uint32_t callId_;
